+++ CHANGELOG.md
@@ -1,8 +1,13 @@
+## 2.1.0
+
+- Add `Markup` class.
+
 ## 2.0.0
 
 - Add `Component` class.
 - Remove `Tag` class.
 - Remove `View` class.
+- Remove `Markup` class.
 - Update comments.
 
 ## 1.0.0